requirements.txt
matplotlib==3.1.1
opencv_python==4.1.1.26
scipy==0.19.1
tensorflow==2.0.0
imageio==2.6.1
numpy==1.17.3
seaborn==0.9.0
requests==2.22.0
Pillow==6.2.1
tensorflow==2.0.0
